// src/api/userProfile/userProfile.enum.ts
export enum genderType {
    Male = 'male',
    Female = 'female',
}

export enum horoscopeType {
    Aries = 'aries',
    Taurus = 'taurus',
    Gemini = 'gemini',
    Cancer = 'cancer',
    Leo = 'leo',
    Virgo = 'virgo',
    Libra = 'libra',
    Scorpius = 'scorpius',
    Sagittarius = 'sagittarius',
    Capricornus = 'capricornus',
    Aquarius = 'aquarius',
    Pisces = 'pisces',
}

export enum zodiacType {
    Rabbit = 'rabbit',
    Tiger = 'tiger',
    Ox = 'ox',
    Rat = 'rat',
    Pig = 'pig',
    Dog = 'dog',
    Rooster = 'rooster',
    Monkey = 'monkey',
    Goat = 'goat',
    Horse = 'horse',
    Snake = 'snake',
    Dragon = 'dragon',
}